requirements.txt
discord.py==2.4.0
python-dotenv==1.0.0
httpx==0.27.0
aiohttp==3.10.0
aiosqlite==3.1.0
